+++ Sources/MintCLI/Commands/PackageCommand.swift
@@ -6,13 +6,14 @@ import SwiftCLI
 class PackageCommand: MintfileCommand {
 
     var package = Parameter()
+    var silent = Flag("-s", "--silent", description: "Silences any output from Mint itself")
 
     init(mint: Mint, name: String, description: String, parameterDescription: String? = nil) {
         var longDescription = """
         \(description)
         
         The package can be a shorthand for a github repo \"githubName/repo\", or a fully qualified .git path.
-        An optional version can be specified by appending @version to the repo, otherwise the newest tag will be used (or master if no tags are found)
+        An optional version can be specified by appending @version to the repo, otherwise the newest tag will be used (or master if no tags are found.)
         """
         if let parameterDescription = parameterDescription {
             longDescription += "\n\n\(parameterDescription)"
@@ -22,23 +23,16 @@ class PackageCommand: MintfileCommand {
     }
 
     override func execute() throws {
-        try super.execute()
-
-        var mintPackage = MintPackage(package: package.value)
-
-        if mintPackage.version.isEmpty,
-            mint.mintFilePath.exists,
-            let mintfile = try? Mintfile(path: mint.mintFilePath) {
-            // set version to version from mintfile
-            if let package = mintfile.package(for: mintPackage.repo), !package.version.isEmpty {
-                mintPackage = package
-                mint.standardOut <<< "🌱  Using \"\(package.repo)\" \"\(package.version)\" from Mintfile."
-            }
+        if silent.value {
+            mint.standardOut = WriteStream.null
         }
 
-        try execute(repo: mintPackage.repo, version: mintPackage.version)
+        try super.execute()
+
+        let package = PackageReference(package: self.package.value)
+        try execute(package: package)
     }
 
-    func execute(repo: String, version: String) throws {
+    func execute(package: PackageReference) throws {
     }
 }