+++ Sources/MintCLI/Commands/RunCommand.swift
@@ -4,32 +4,16 @@ import SwiftCLI
 
 class RunCommand: PackageCommand {
 
-    var command = OptionalCollectedParameter()
-    var silent = Flag("-s", "--silent", description: "Silences any output from Mint itself")
+    var arguments = OptionalCollectedParameter()
 
     init(mint: Mint) {
         super.init(mint: mint,
                    name: "run",
-                   description: "Install and then run a package")
+                   description: "Install and then run a package",
+                   parameterDescription: "The arguments can be used to specify a specific executable and it's arguments. By default the single executable in the Package.swift will be used, otherwise if there are multiple it will ask you to choose")
     }
 
-    override func execute() throws {
-        if silent.value {
-            mint.standardOut = LineStream {_ in}
-        }
-        try super.execute()
-    }
-
-    override func execute(repo: String, version: String) throws {
-        var arguments = command.value
-
-        // backwards compatability for arguments surrounded in quotes
-        if arguments.count == 1,
-            let firstArg = arguments.first,
-            firstArg.contains(" ") {
-            arguments = firstArg.split(separator: " ").map(String.init)
-        }
-
-        try mint.run(repo: repo, version: version, arguments: arguments)
+    override func execute(package: PackageReference) throws {
+        try mint.run(package: package, arguments: arguments.value)
     }
 }