+++ Sources/MintKit/Mint.swift
@@ -7,7 +7,7 @@ import Utility
 public class Mint {
 
     public var path: Path
-    public var installationPath: Path
+    public var linkPath: Path
     public var mintFilePath: Path
 
     var packagesPath: Path {
@@ -23,18 +23,20 @@ public class Mint {
 
     public var verbose = false
     public var runAsNewProcess = true
+    var inputReader: InputReader
 
     public init(
         path: Path,
-        installationPath: Path,
+        linkPath: Path,
         mintFilePath: Path = "Mintfile",
         standardOut: WritableStream = WriteStream.stdout,
         standardError: WritableStream = WriteStream.stderr) {
         self.standardOut = standardOut
         self.standardError = standardError
         self.path = path.absolute()
-        self.installationPath = installationPath.absolute()
+        self.linkPath = linkPath.absolute()
         self.mintFilePath = mintFilePath
+        inputReader = InputReader(standardOut: standardOut)
     }
 
     public func closeStreams() {
@@ -70,9 +72,9 @@ public class Mint {
         return metadata.packages.first(where: { $0.key.lowercased().contains(name.lowercased()) })?.key
     }
 
-    func getGlobalInstalledPackages() -> [String: String] {
-        guard installationPath.exists,
-            let packages = try? installationPath.children() else {
+    func getLinkedPackages() -> [String: String] {
+        guard linkPath.exists,
+            let packages = try? linkPath.children() else {
             return [:]
         }
 
@@ -94,7 +96,7 @@ public class Mint {
             return [:]
         }
 
-        let globalInstalledPackages: [String: String] = getGlobalInstalledPackages()
+        let linkedPackages: [String: String] = getLinkedPackages()
 
         var versionsByPackage: [String: [String]] = [:]
         let packages: [String] = try packagesPath.children().filter { $0.isDirectory }.map { packagePath in
@@ -107,73 +109,44 @@ public class Mint {
             var package = "  \(packageName)"
             for version in versions {
                 package += "\n    - \(version)"
-                if globalInstalledPackages[packageName] == version {
+                if linkedPackages[packageName] == version {
                     package += " *"
                 }
                 versionsByPackage[packageName, default: []].append(version)
             }
             return package
-        }
+        }.sorted { $0.localizedStandardCompare($1) == .orderedAscending }
 
-        standardOut <<< "Installed mint packages:\n\(packages.sorted().joined(separator: "\n"))"
+        standardOut <<< "Installed mint packages:\n\(packages.joined(separator: "\n"))"
         return versionsByPackage
     }
 
-    @discardableResult
-    public func run(repo: String, version: String, arguments: [String]? = nil) throws -> Package {
-        let guessedCommand = repo.components(separatedBy: "/").last!.components(separatedBy: ".").first!
-        let name = arguments?.first ?? guessedCommand
-        var arguments = arguments ?? [guessedCommand]
-        arguments = arguments.count > 1 ? Array(arguments.dropFirst()) : []
-        var git = repo
-        if !git.contains("/") {
-            // find repo
-            if let existingGit = try getPackageGit(name: git) {
-                git = existingGit
-            } else {
-                throw MintError.packageNotFound(git)
+    func resolvePackage(_ package: PackageReference) throws {
+
+        // resolve version from MintFile
+        if package.version.isEmpty,
+            mintFilePath.exists,
+            let mintfile = try? Mintfile(path: mintFilePath) {
+            // set version to version from mintfile
+            if let mintFilePackage = mintfile.package(for: package.repo), !mintFilePackage.version.isEmpty {
+                package.version = mintFilePackage.version
+                package.repo = mintFilePackage.repo
+                standardOut <<< "🌱  Using \(package.repo) \(package.version) from Mintfile."
             }
         }
-        let package = Package(repo: git, version: version, name: name)
-        try run(package, arguments: arguments, verbose: verbose)
-        return package
-    }
 
-    public func run(_ package: Package, arguments: [String], verbose: Bool) throws {
-        try install(package, update: false, global: false)
-        standardOut <<< "🌱  Running \(package.commandVersion)..."
-        let packagePath = PackagePath(path: packagesPath, package: package)
-        if !packagePath.commandPath.exists {
-            throw MintError.invalidCommand(packagePath.commandPath.string)
-        }
-        if runAsNewProcess {
-            var env = ProcessInfo.processInfo.environment
-            env["MINT"] = "YES"
-            env["RESOURCE_PATH"] = ""
-            try Task.execvp(packagePath.commandPath.string, arguments: arguments, env: env)
-        } else {
-            let runTask = Task(executable: packagePath.commandPath.string, arguments: arguments)
-            _ = runTask.runSync()
-        }
-    }
-
-    @discardableResult
-    public func install(repo: String, version: String, command: String?, update: Bool = false, global: Bool = false) throws -> Package {
-        let guessedCommand = repo.components(separatedBy: "/").last!.components(separatedBy: ".").first!
-        let name = command ?? guessedCommand
-        let package = Package(repo: repo, version: version, name: name)
-        try install(package, update: update, global: global)
-        return package
-    }
-
-    public func install(_ package: Package, update: Bool = false, global: Bool = false) throws {
-
+        // resolve repo from installed packages
         if !package.repo.contains("/") {
-            throw MintError.invalidRepo(package.repo)
+            // repo reference by name. Get the full git repo
+            if let existingGit = try getPackageGit(name: package.repo) {
+                package.repo = existingGit
+            } else {
+                throw MintError.packageNotFound(package.repo)
+            }
         }
 
+        // resove latest version from git repo
         let packagePath = PackagePath(path: packagesPath, package: package)
-
         if package.version.isEmpty {
             // we don't have a specific version, let's get the latest tag
             standardOut <<< "🌱  Finding latest version of \(package.name)"
@@ -198,12 +171,66 @@ public class Mint {
                 throw MintError.repoNotFound(packagePath.gitPath)
             }
         }
+    }
 
-        if !update && packagePath.commandPath.exists {
-            if global {
-                try installGlobal(packagePath: packagePath)
-            } else {
-                standardOut <<< "🌱  \(package.commandVersion) already installed".green
+    public func run(package: PackageReference, arguments: [String] = []) throws {
+
+        try resolvePackage(package)
+
+        // install the package if not installed already
+        try install(package: package, force: false, link: false)
+
+        var packagePath = PackagePath(path: packagesPath, package: package)
+
+        if let packageExecutable = arguments.first {
+            packagePath.executable = packageExecutable
+            if !packagePath.executablePath.exists {
+                throw MintError.invalidExecutable(packageExecutable)
+            }
+        } else {
+            let executables = try packagePath.getExecutables()
+            switch executables.count {
+            case 0:
+                throw MintError.missingExecutable
+            case 1:
+                packagePath.executable = executables[0]
+            default:
+                packagePath.executable = inputReader.ask("There are multiple executables, which one would you like to run?", answers: executables)
+            }
+        }
+        standardOut <<< "🌱  Running \(packagePath.executable ?? "") \(package.version)..."
+
+        let arguments = arguments.isEmpty ? [] : Array(arguments.dropFirst())
+
+        if runAsNewProcess {
+            var env = ProcessInfo.processInfo.environment
+            env["MINT"] = "YES"
+            env["RESOURCE_PATH"] = ""
+            try Task.execvp(packagePath.executablePath.string, arguments: arguments, env: env)
+        } else {
+            let runTask = Task(executable: packagePath.executablePath.string, arguments: arguments)
+            _ = runTask.runSync()
+        }
+    }
+
+    public func install(package: PackageReference, executable: String? = nil, force: Bool = false, link: Bool = false) throws {
+
+        try resolvePackage(package)
+
+        let packagePath = PackagePath(path: packagesPath, package: package, executable: executable)
+
+        let alreadyInstalled = packagePath.installPath.exists
+        if !force && alreadyInstalled {
+            standardOut <<< "🌱  \(packagePath.commandVersion) already installed".green
+            if link {
+                if let executable = executable {
+                    try linkPackage(package, executable: executable)
+                } else {
+                    let executables = try packagePath.getExecutables()
+                    for executable in executables {
+                        try linkPackage(package, executable: executable)
+                    }
+                }
             }
             return
         }
@@ -233,23 +260,34 @@ public class Mint {
             throw MintError.cloneError(url: packagePath.gitPath, version: package.version)
         }
 
-        standardOut <<< "🌱  Building \(package.name) with SPM..."
+        guard let spmPackage = try? SwiftPackage(directory: packageCheckoutPath) else {
+            throw MintError.packageFileNotFound
+        }
+
+        let executables = spmPackage.products.filter { $0.isExecutable }.map { $0.name }
+        guard !executables.isEmpty else {
+            throw MintError.missingExecutable
+        }
+
+        standardOut <<< "🌱  Building \(spmPackage.name) Package with SPM..."
 
         try buildPackage(name: package.name, path: packageCheckoutPath)
 
-        standardOut <<< "🌱  Installing..."
+        standardOut <<< "🌱  Installing \(spmPackage.name)..."
 
-        let toolFile = packageCheckoutPath + ".build/release/\(package.name)"
-        if !toolFile.exists {
-            throw MintError.invalidCommand(package.name)
-        }
-
-        // TODO: perhaps don't remove the whole directory once we install specific executables
+        // clear the install directory
         try? packagePath.installPath.delete()
         try packagePath.installPath.mkpath()
 
-        // copy using shell instead of FileManager via PathKit because it remove executable permissions on Linux
-        try SwiftCLI.run("cp", toolFile.string, packagePath.commandPath.string)
+        for executable in executables {
+            let executablePath = packageCheckoutPath + ".build/release/\(executable)"
+            if !executablePath.exists {
+                throw MintError.invalidExecutable(executablePath.lastComponent)
+            }
+            let destinationPackagePath = PackagePath(path: packagesPath, package: package, executable: executable)
+            // copy using shell instead of FileManager via PathKit because it removes executable permissions on Linux
+            try SwiftCLI.run("cp", executablePath.string, destinationPackagePath.executablePath.string)
+        }
 
         let resourcesFile = packageCheckoutPath + "Package.resources"
         if resourcesFile.exists {
@@ -257,11 +295,13 @@ public class Mint {
             let resources = resourcesString.components(separatedBy: "\n")
                 .map { $0.trimmingCharacters(in: .whitespacesAndNewlines) }
                 .filter { !$0.isEmpty }
-            standardOut <<< "🌱  Copying resources for \(package.name): \(resources.joined(separator: ", ")) ..."
+            standardOut <<< "🌱  Copying resources for \(spmPackage.name): \(resources.joined(separator: ", ")) ..."
             for resource in resources {
                 let resourcePath = packageCheckoutPath + resource
                 if resourcePath.exists {
-                    try resourcePath.copy(packagePath.installPath + resource)
+                    let filename = String(resource.split(separator: "/").last!)
+                    let dest = packagePath.installPath + filename
+                    try SwiftCLI.run(bash: "cp -R \"\(resourcePath)\" \"\(dest)\"")
                 } else {
                     standardOut <<< "resource \(resource) doesn't exist".yellow
                 }
@@ -272,12 +312,18 @@ public class Mint {
 
         try addPackage(git: packagePath.gitPath, path: packagePath.packagePath)
 
-        standardOut <<< "🌱  Installed \(package.commandVersion)".green
-        if global {
-            try installGlobal(packagePath: packagePath)
-        }
-
+        standardOut <<< "🌱  Installed \(spmPackage.name) \(package.version)".green
         try? packageCheckoutPath.delete()
+
+        if link {
+            if let executable = executable {
+                try linkPackage(package, executable: executable)
+            } else {
+                for executable in executables {
+                    try linkPackage(package, executable: executable)
+                }
+            }
+        }
     }
 
     private func installManPages(packagePath: PackagePath) throws {
@@ -343,8 +389,8 @@ public class Mint {
 //            buildStepOutput += string + "\n"
 //        }
 
-        let taskOut = verbose ? standardOut : LineStream {_ in}
-        let taskError = LineStream {_ in}
+        let taskOut = verbose ? standardOut : WriteStream.null
+        let taskError = WriteStream.null
         let task = Task(executable: "/bin/bash", arguments: ["-c", command], directory: path.string, stdout: taskOut, stderr: taskError)
         task.runAsync()
         let status = task.finish()
@@ -379,7 +425,7 @@ public class Mint {
         }()
 
         if let warning = warning {
-            let ok = Question().confirmation("🌱  \(warning)\nOvewrite it with Mint's symlink?".yellow)
+            let ok = InputReader().confirmation("🌱  \(warning)\nOvewrite it with Mint's symlink?".yellow)
             if !ok {
                 return false
             }
@@ -388,10 +434,10 @@ public class Mint {
         return true
     }
 
-    func installGlobal(packagePath: PackagePath) throws {
+    func linkPackage(_ package: PackageReference, executable: String) throws {
 
-        let toolPath = packagePath.commandPath
-        let installPath = installationPath + packagePath.package.name
+        let packagePath = PackagePath(path: packagesPath, package: package, executable: executable)
+        let installPath = linkPath + packagePath.executable!
 
         let installStatus = try InstallStatus(path: installPath, mintPackagesPath: packagesPath)
 
@@ -403,12 +449,12 @@ public class Mint {
         try? installPath.parent().mkpath()
 
         do {
-            try SwiftCLI.run(bash: "ln -s \(toolPath.string) \(installPath.string)")
+            try SwiftCLI.run(bash: "ln -s \(packagePath.executablePath.string) \(installPath.string)")
         } catch {
-            standardError <<< "🌱  Could not install \(packagePath.package.commandVersion) in \(installPath.string)"
+            standardError <<< "🌱  Could not link \(packagePath.commandVersion) to \(installPath.string)"
             return
         }
-        var confirmation = "Linked \(packagePath.package.commandVersion) to \(installationPath.string)"
+        var confirmation = "Linked \(packagePath.commandVersion) to \(linkPath.string)"
         if case let .mint(previousVersion) = installStatus.status {
             confirmation += ", replacing version \(previousVersion)"
         }
@@ -438,7 +484,7 @@ public class Mint {
                 }()
 
                 if let warning = warning {
-                    let ok = Question().confirmation("🌱  \(warning)\nOvewrite it with Mint's symlink?".yellow)
+                    let ok = InputReader().confirmation("🌱  \(warning)\nOvewrite it with Mint's symlink?".yellow)
                     if !ok {
                         continue
                     }
@@ -462,8 +508,8 @@ public class Mint {
         let packageCount = "\(mintFile.packages.count) \(mintFile.packages.count == 1 ? "package" : "packages")"
 
         standardOut <<< "🌱  Found \(packageCount) in \(mintFilePath.string)"
-        for mintPackage in mintFile.packages {
-            try install(repo: mintPackage.repo, version: mintPackage.version, command: nil, update: false, global: false)
+        for package in mintFile.packages {
+            try install(package: package, force: false, link: false)
         }
         standardOut <<< "🌱  Installed \(packageCount) from \(mintFilePath.string)".green
     }
@@ -499,8 +545,8 @@ public class Mint {
         }
         try writeMetadata(metadata)
 
-        // remove global install
-        let installPath = installationPath + name
+        // remove link
+        let installPath = linkPath + name
 
         let installStatus = try InstallStatus(path: installPath, mintPackagesPath: packagesPath)