+++ Tests/MintTests/XCTestManifests.swift
@@ -2,11 +2,10 @@ import XCTest
 
 extension MintTests {
     static let __allTests = [
+        ("testBootstrapCommand", testBootstrapCommand),
         ("testInstallCommand", testInstallCommand),
         ("testMintErrors", testMintErrors),
-        ("testMintPackageInfo", testMintPackageInfo),
-        ("testPackageGitPaths", testPackageGitPaths),
-        ("testPackagePaths", testPackagePaths),
+        ("testMintFileInstall", testMintFileInstall),
         ("testRunCommand", testRunCommand),
     ]
 }
@@ -18,11 +17,20 @@ extension MintfileTests {
     ]
 }
 
-#if !os(macOS)
-public func __allTests() -> [XCTestCaseEntry] {
-    return [
-        testCase(MintTests.__allTests),
-        testCase(MintfileTests.__allTests),
+extension PackageTests {
+    static let __allTests = [
+        ("testPackageGitPaths", testPackageGitPaths),
+        ("testPackagePaths", testPackagePaths),
+        ("testPackageReferenceInfo", testPackageReferenceInfo),
     ]
 }
+
+#if !os(macOS)
+    public func __allTests() -> [XCTestCaseEntry] {
+        return [
+            testCase(MintTests.__allTests),
+            testCase(MintfileTests.__allTests),
+            testCase(PackageTests.__allTests),
+        ]
+    }
 #endif